refactor: socket 로직에서 게임 로직 분리 끝
