+++ src/game/maps/index.ts
@@ -1,3 +1,3 @@
 import { CommonMap } from './common'
 import { TailTagMap } from './tail-tag'
-export { CommonMap, TailTagMap as TailTagRound }
+export { CommonMap, TailTagMap }
