+++ src/game/maps/tail-tag.ts
@@ -4,6 +4,12 @@ import { CommonMap } from './common'
 const TAIL_STEAL_DISTANCE = 5
 
 export class TailTagMap extends CommonMap {
+    private calculateDistance(pos1: Position, pos2: Position) {
+        const dx = pos1[0] - pos2[0]
+        const dz = pos1[2] - pos2[2]
+        return Math.sqrt(dx * dx + dz * dz)
+    }
+
     handleCatch(character: Character) {
         if (character.hasTail) return
 
@@ -22,18 +28,18 @@ export class TailTagMap extends CommonMap {
                     other.isBeingStolen = true
                     character.hasTail = true
                     other.hasTail = false
-                    console.log(
-                        `${character.id} has stolen the tail from ${other.id}`
-                    )
                     break
                 }
             }
         }
     }
 
-    private calculateDistance(pos1: Position, pos2: Position) {
-        const dx = pos1[0] - pos2[0]
-        const dz = pos1[2] - pos2[2]
-        return Math.sqrt(dx * dx + dz * dz)
+    updateGameState(): void {
+        super.updateGameState()
+
+        this.characters.forEach((character) => {
+            character.updatePosition()
+            character.isBeingStolen = false
+        })
     }
 }
